requirements.txt
django==2.0.5
djangorestframework==3.8.2
drf-nested-routers==0.90.2
